Updater: Remove config_hideRecoveryUpdate and "Update recovery" on AB
Devices with prebuilt vendor are not taken into account because
nobody does that anymore. The "Update recovery" menu preference
visibility is now opt-in and is meant only for the deprecated
Non-AB OTA methodology.

The same menu is now not visible in AB devices and thus the toast
message is also removed because the users are now fairly adapted to
the recovery being updated and slightly better UX to avoid a greyed
out preference button.

Change-Id: I0973de60530d0800a93e2b107996d78cffd8649e